+++ src/error_handling/mod.rs
@@ -203,7 +203,7 @@ pub mod future {
 
 #[test]
 fn traits() {
-    use crate::tests::*;
+    use crate::test_helpers::*;
 
     assert_send::<HandleError<(), (), NotSendSync>>();
     assert_sync::<HandleError<(), (), NotSendSync>>();
