+++ src/routing/route.rs
@@ -95,7 +95,7 @@ mod tests {
 
     #[test]
     fn traits() {
-        use crate::tests::*;
+        use crate::test_helpers::*;
         assert_send::<Route<()>>();
     }
 }
