+++ src/routing/service_method_routing.rs
@@ -553,7 +553,7 @@ where
 
 #[test]
 fn traits() {
-    use crate::tests::*;
+    use crate::test_helpers::*;
 
     assert_send::<MethodRouter<(), (), NotSendSync>>();
     assert_sync::<MethodRouter<(), (), NotSendSync>>();
