+++ src/routing/tests/mod.rs
@@ -1,7 +1,6 @@
-#![allow(clippy::blacklisted_name)]
-
 use crate::error_handling::HandleErrorLayer;
 use crate::extract::{Extension, MatchedPath};
+use crate::test_helpers::*;
 use crate::BoxError;
 use crate::{
     extract::{self, Path},
@@ -31,12 +30,9 @@ use tower::{service_fn, timeout::TimeoutLayer, ServiceBuilder, ServiceExt};
 use tower_http::auth::RequireAuthorizationLayer;
 use tower_service::Service;
 
-pub(crate) use helpers::*;
-
 mod fallback;
 mod get_to_head;
 mod handle_error;
-mod helpers;
 mod merge;
 mod nest;
 
@@ -71,105 +67,6 @@ async fn hello_world() {
     let res = client.post("/users").send().await;
     let body = res.text().await;
     assert_eq!(body, "users#create");
-}
-
-#[tokio::test]
-async fn consume_body() {
-    let app = Router::new().route("/", get(|body: String| async { body }));
-
-    let client = TestClient::new(app);
-    let res = client.get("/").body("foo").send().await;
-    let body = res.text().await;
-
-    assert_eq!(body, "foo");
-}
-
-#[tokio::test]
-async fn deserialize_body() {
-    #[derive(Debug, Deserialize)]
-    struct Input {
-        foo: String,
-    }
-
-    let app = Router::new().route(
-        "/",
-        post(|input: extract::Json<Input>| async { input.0.foo }),
-    );
-
-    let client = TestClient::new(app);
-    let res = client.post("/").json(&json!({ "foo": "bar" })).send().await;
-    let body = res.text().await;
-
-    assert_eq!(body, "bar");
-}
-
-#[tokio::test]
-async fn consume_body_to_json_requires_json_content_type() {
-    #[derive(Debug, Deserialize)]
-    struct Input {
-        foo: String,
-    }
-
-    let app = Router::new().route(
-        "/",
-        post(|input: extract::Json<Input>| async { input.0.foo }),
-    );
-
-    let client = TestClient::new(app);
-    let res = client.post("/").body(r#"{ "foo": "bar" }"#).send().await;
-
-    let status = res.status();
-    dbg!(res.text().await);
-
-    assert_eq!(status, StatusCode::BAD_REQUEST);
-}
-
-#[tokio::test]
-async fn body_with_length_limit() {
-    use std::iter::repeat;
-
-    #[derive(Debug, Deserialize)]
-    struct Input {
-        foo: String,
-    }
-
-    const LIMIT: u64 = 8;
-
-    let app = Router::new().route(
-        "/",
-        post(|_body: extract::ContentLengthLimit<Bytes, LIMIT>| async {}),
-    );
-
-    let client = TestClient::new(app);
-    let res = client
-        .post("/")
-        .body(repeat(0_u8).take((LIMIT - 1) as usize).collect::<Vec<_>>())
-        .send()
-        .await;
-    assert_eq!(res.status(), StatusCode::OK);
-
-    let res = client
-        .post("/")
-        .body(repeat(0_u8).take(LIMIT as usize).collect::<Vec<_>>())
-        .send()
-        .await;
-    assert_eq!(res.status(), StatusCode::OK);
-
-    let res = client
-        .post("/")
-        .body(repeat(0_u8).take((LIMIT + 1) as usize).collect::<Vec<_>>())
-        .send()
-        .await;
-    assert_eq!(res.status(), StatusCode::PAYLOAD_TOO_LARGE);
-
-    let res = client
-        .post("/")
-        .body(reqwest::Body::wrap_stream(futures_util::stream::iter(
-            vec![Ok::<_, std::io::Error>(bytes::Bytes::new())],
-        )))
-        .send()
-        .await;
-    assert_eq!(res.status(), StatusCode::LENGTH_REQUIRED);
 }
 
 #[tokio::test]
@@ -209,42 +106,8 @@ async fn routing() {
 }
 
 #[tokio::test]
-async fn extracting_url_params() {
-    let app = Router::new().route(
-        "/users/:id",
-        get(|Path(id): Path<i32>| async move {
-            assert_eq!(id, 42);
-        })
-        .post(|Path(params_map): Path<HashMap<String, i32>>| async move {
-            assert_eq!(params_map.get("id").unwrap(), &1337);
-        }),
-    );
-
-    let client = TestClient::new(app);
-
-    let res = client.get("/users/42").send().await;
-    assert_eq!(res.status(), StatusCode::OK);
-
-    let res = client.post("/users/1337").send().await;
-    assert_eq!(res.status(), StatusCode::OK);
-}
-
-#[tokio::test]
-async fn extracting_url_params_multiple_times() {
-    let app = Router::new().route(
-        "/users/:id",
-        get(|_: extract::Path<i32>, _: extract::Path<String>| async {}),
-    );
-
-    let client = TestClient::new(app);
-
-    let res = client.get("/users/42").send().await;
-    assert_eq!(res.status(), StatusCode::OK);
-}
-
-#[tokio::test]
-async fn boxing() {
-    let app = Router::new()
+async fn router_type_doesnt_change() {
+    let app: Router = Router::new()
         .route(
             "/",
             on(MethodFilter::GET, |_: Request<Body>| async {
@@ -352,49 +215,6 @@ async fn service_in_bottom() {
 }
 
 #[tokio::test]
-async fn test_extractor_middleware() {
-    struct RequireAuth;
-
-    #[async_trait::async_trait]
-    impl<B> extract::FromRequest<B> for RequireAuth
-    where
-        B: Send,
-    {
-        type Rejection = StatusCode;
-
-        async fn from_request(req: &mut extract::RequestParts<B>) -> Result<Self, Self::Rejection> {
-            if let Some(auth) = req
-                .headers()
-                .expect("headers already extracted")
-                .get("authorization")
-                .and_then(|v| v.to_str().ok())
-            {
-                if auth == "secret" {
-                    return Ok(Self);
-                }
-            }
-
-            Err(StatusCode::UNAUTHORIZED)
-        }
-    }
-
-    async fn handler() {}
-
-    let app = Router::new().route(
-        "/",
-        get(handler.layer(extract::extractor_middleware::<RequireAuth>())),
-    );
-
-    let client = TestClient::new(app);
-
-    let res = client.get("/").send().await;
-    assert_eq!(res.status(), StatusCode::UNAUTHORIZED);
-
-    let res = client.get("/").header(AUTHORIZATION, "secret").send().await;
-    assert_eq!(res.status(), StatusCode::OK);
-}
-
-#[tokio::test]
 async fn wrong_method_handler() {
     let app = Router::new()
         .route("/", get(|| async {}).post(|| async {}))
@@ -468,56 +288,6 @@ async fn multiple_methods_for_one_handler() {
 
     let res = client.post("/").send().await;
     assert_eq!(res.status(), StatusCode::OK);
-}
-
-#[tokio::test]
-async fn handler_into_service() {
-    async fn handle(body: String) -> impl IntoResponse {
-        format!("you said: {}", body)
-    }
-
-    let client = TestClient::new(handle.into_service());
-
-    let res = client.post("/").body("hi there!").send().await;
-    assert_eq!(res.status(), StatusCode::OK);
-    assert_eq!(res.text().await, "you said: hi there!");
-}
-
-#[tokio::test]
-async fn captures_dont_match_empty_segments() {
-    let app = Router::new().route("/:key", get(|| async {}));
-
-    let client = TestClient::new(app);
-
-    let res = client.get("/").send().await;
-    assert_eq!(res.status(), StatusCode::NOT_FOUND);
-
-    let res = client.get("/foo").send().await;
-    assert_eq!(res.status(), StatusCode::OK);
-}
-
-#[tokio::test]
-async fn json_content_types() {
-    async fn valid_json_content_type(content_type: &str) -> bool {
-        println!("testing {:?}", content_type);
-
-        let app = Router::new().route("/", post(|Json(_): Json<Value>| async {}));
-
-        let res = TestClient::new(app)
-            .post("/")
-            .header("content-type", content_type)
-            .body("{}")
-            .send()
-            .await;
-
-        res.status() == StatusCode::OK
-    }
-
-    assert!(valid_json_content_type("application/json").await);
-    assert!(valid_json_content_type("application/json; charset=utf-8").await);
-    assert!(valid_json_content_type("application/json;charset=utf-8").await);
-    assert!(valid_json_content_type("application/cloudevents+json").await);
-    assert!(!valid_json_content_type("text/json").await);
 }
 
 #[tokio::test]
@@ -634,94 +404,6 @@ async fn wildcard_with_trailing_slash() {
     );
 }
 
-#[derive(Clone)]
-struct SetMatchedPathExtension<S>(S);
-
-impl<B, S> Service<Request<B>> for SetMatchedPathExtension<S>
-where
-    S: Service<Request<B>>,
-{
-    type Response = S::Response;
-    type Error = S::Error;
-    type Future = S::Future;
-
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        self.0.poll_ready(cx)
-    }
-
-    fn call(&mut self, mut req: Request<B>) -> Self::Future {
-        let path = req
-            .extensions()
-            .get::<MatchedPath>()
-            .unwrap()
-            .as_str()
-            .to_string();
-        req.extensions_mut().insert(MatchedPathFromMiddleware(path));
-        self.0.call(req)
-    }
-}
-
-#[derive(Clone)]
-struct MatchedPathFromMiddleware(String);
-
-#[tokio::test]
-async fn access_matched_path() {
-    let api = Router::new().route(
-        "/users/:id",
-        get(|path: MatchedPath| async move { path.as_str().to_string() }),
-    );
-
-    async fn handler(
-        path: MatchedPath,
-        Extension(MatchedPathFromMiddleware(path_from_middleware)): Extension<
-            MatchedPathFromMiddleware,
-        >,
-    ) -> String {
-        format!(
-            "extractor = {}, middleware = {}",
-            path.as_str(),
-            path_from_middleware
-        )
-    }
-
-    let app = Router::new()
-        .route(
-            "/:key",
-            get(|path: MatchedPath| async move { path.as_str().to_string() }),
-        )
-        .nest("/api", api)
-        .nest(
-            "/public",
-            Router::new().route("/assets/*path", get(handler)),
-        )
-        .nest("/foo", handler.into_service())
-        .layer(tower::layer::layer_fn(SetMatchedPathExtension));
-
-    let client = TestClient::new(app);
-
-    let res = client.get("/foo").send().await;
-    assert_eq!(res.text().await, "/:key");
-
-    let res = client.get("/api/users/123").send().await;
-    assert_eq!(res.text().await, "/api/users/:id");
-
-    let res = client.get("/public/assets/css/style.css").send().await;
-    assert_eq!(
-        res.text().await,
-        "extractor = /public/assets/*path, middleware = /public/assets/*path"
-    );
-
-    let res = client.get("/foo/bar/baz").send().await;
-    assert_eq!(
-        res.text().await,
-        format!(
-            "extractor = /foo/*{}, middleware = /foo/*{}",
-            crate::routing::NEST_TAIL_PARAM,
-            crate::routing::NEST_TAIL_PARAM,
-        ),
-    );
-}
-
 #[tokio::test]
 async fn static_and_dynamic_paths() {
     let app = Router::new()
@@ -801,9 +483,3 @@ async fn middleware_still_run_for_unmatched_requests() {
 async fn routing_to_router_panics() {
     TestClient::new(Router::new().route("/", Router::new()));
 }
-
-pub(crate) fn assert_send<T: Send>() {}
-pub(crate) fn assert_sync<T: Sync>() {}
-pub(crate) fn assert_unpin<T: Unpin>() {}
-
-pub(crate) struct NotSendSync(*const ());
